# tiki-backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vona API",
    "version": "1.0.0",
    "description": "API documentation for Vona application"
  },
  "servers": [
    {
      "url": "http://localhost:3003/api/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "summary": "Get list of users",
        "description": "Retrieve a paginated list of users with optional filters (requires user login)",
        "operationId": "getAllUsers",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filter parameters",
            "schema": { "$ref": "#/components/schemas/UserFilter" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user with provided data",
        "operationId": "createUser",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "summary": "Get list of products",
        "description": "Retrieve a paginated list of products with optional filters",
        "operationId": "getAllProducts",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filter parameters for products",
            "schema": { "$ref": "#/components/schemas/ProductFilter" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "description": "Create a new product with provided data (requires admin login)",
        "operationId": "createProduct",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/addresses": {
      "get": {
        "summary": "Get list of addresses",
        "description": "Retrieve a list of addresses for the authenticated user",
        "operationId": "getAllAddresses",
        "tags": ["Addresses"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of addresses",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new address",
        "description": "Create a new address for the authenticated user",
        "operationId": "createAddress",
        "tags": ["Addresses"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddressCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddressResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/rate-limit": {
      "get": {
        "summary": "Get rate limit information",
        "description": "Retrieve rate limit information for the API",
        "operationId": "getRateLimit",
        "tags": ["RateLimit"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Rate limit information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/permissions": {
      "get": {
        "summary": "Get list of permissions",
        "description": "Retrieve a list of permissions (requires MANAGE_PERMISSIONS)",
        "operationId": "getAllPermissions",
        "tags": ["Permissions"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of permissions",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get list of orders",
        "description": "Retrieve a paginated list of orders for the authenticated user",
        "operationId": "getAllOrders",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "description": "Create a new order with provided data",
        "operationId": "createOrder",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/categories": {
      "get": {
        "summary": "Get list of categories",
        "description": "Retrieve a paginated list of categories",
        "operationId": "getAllCategories",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of categories to return"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of categories to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new category",
        "description": "Create a new category (requires MANAGE_CATEGORIES)",
        "operationId": "createCategory",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoryResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/cart-items": {
      "post": {
        "summary": "Add item to cart",
        "description": "Add a new item to the cart",
        "operationId": "createCartItem",
        "tags": ["CartItems"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CartItemCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cart item added successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CartItemResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/chatbot": {
      "post": {
        "summary": "Interact with chatbot",
        "description": "Send a message to the chatbot and receive a response",
        "operationId": "interactWithChatbot",
        "tags": ["Chatbot"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "minLength": 1 }
                },
                "required": ["message"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chatbot response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/reviews": {
      "get": {
        "summary": "Get list of reviews",
        "description": "Retrieve a paginated list of reviews with optional filters",
        "operationId": "getAllReviews",
        "tags": ["Reviews"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filter parameters",
            "schema": { "$ref": "#/components/schemas/ReviewFilter" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new review",
        "description": "Create a new review for a product (requires CREATE_REVIEW or ownership)",
        "operationId": "createReview",
        "tags": ["Reviews"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReviewCreateInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReviewResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "full_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email" },
          "phone_number": { "type": "string", "nullable": true },
          "day_of_birth": { "type": "string", "nullable": true },
          "is_public": { "type": "boolean" },
          "is_deleted": { "type": "boolean" },
          "gender": { "type": "boolean", "nullable": true },
          "role_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "avatar": { "type": "string", "nullable": true },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "addresses": { "type": "array", "items": {} },
          "orders": { "type": "array", "items": {} },
          "cart_items": { "type": "array", "items": {} },
          "reviews": { "type": "array", "items": {} },
          "conversations": { "type": "array", "items": {} }
        },
        "required": [
          "uuid",
          "username",
          "full_name",
          "email",
          "phone_number",
          "day_of_birth",
          "is_public",
          "is_deleted",
          "gender",
          "role_id",
          "avatar",
          "created_at",
          "updated_at"
        ]
      },
      "UserCreateInput": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "minLength": 3 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "full_name": { "type": "string", "nullable": true },
          "phone_number": { "type": "string", "nullable": true },
          "day_of_birth": { "type": "string", "nullable": true },
          "gender": { "type": "boolean" },
          "avatar": { "type": "string" },
          "role_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "default": 1
          },
          "is_public": { "type": "boolean", "default": true },
          "is_activated": { "type": "boolean", "default": true }
        },
        "required": [
          "username",
          "email",
          "password",
          "full_name",
          "phone_number",
          "day_of_birth",
          "role_id",
          "is_public",
          "is_activated"
        ]
      },
...
      "UserFilter": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string", "minLength": 3 },
          "full_name": { "type": "string" },
          "phone_number": { "type": "string" },
          "sort_by_name": { "type": "string", "enum": ["asc", "desc"] },
          "sort_by_date": { "type": "string", "enum": ["newest", "oldest"] },
          "limit": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "offset": { "type": "integer", "minimum": 0 }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "data": { "nullable": true },
          "statusCode": { "type": "number" }
        },
        "required": ["error", "statusCode"]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "data": {},
          "statusCode": { "type": "number" }
        },
        "required": ["message", "statusCode"]
      },
...
      "ProductResponse": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "price": { "type": "number" },
          "stock": { "type": "number" },
          "category_id": { "type": "number" },
          "is_deleted": { "type": "boolean" },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": { "type": "string", "format": "uri" },
                "is_primary": { "type": "boolean", "default": false }
              },
              "required": ["url", "is_primary"]
            }
          },
          "specifications": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": { "type": "string", "minLength": 1 },
                "value": { "type": "string", "minLength": 1 }
              },
              "required": ["key", "value"]
            }
          },
          "attributes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "minLength": 1 },
                "value": { "type": "string", "minLength": 1 }
              },
              "required": ["name", "value"]
            }
          },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          }
        },
        "required": [
          "uuid",
          "name",
          "description",
          "price",
          "stock",
          "category_id",
          "is_deleted",
          "created_at",
          "updated_at"
        ]
      },
      "ProductCreateInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "description": { "type": "string" },
          "price": { "type": "number", "minimum": 0, "exclusiveMinimum": true },
          "stock": { "type": "integer", "minimum": 0 },
          "category_id": { "type": "string", "format": "uuid" },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": { "type": "string", "format": "uri" },
                "is_primary": { "type": "boolean", "default": false }
              },
              "required": ["url", "is_primary"]
            },
            "default": []
          },
          "specifications": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": { "type": "string", "minLength": 1 },
                "value": { "type": "string", "minLength": 1 }
              },
              "required": ["key", "value"]
            },
            "default": []
          },
          "attributes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "minLength": 1 },
                "value": { "type": "string", "minLength": 1 }
              },
              "required": ["name", "value"]
            },
            "default": []
          }
        },
        "required": ["name", "price", "stock", "category_id", "images", "specifications", "attributes"]
      },
...
      "ProductFilter": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "category_id": { "type": "string", "format": "uuid" },
          "min_price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "max_price": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "sort_by_price": { "type": "string", "enum": ["asc", "desc"] },
          "limit": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "offset": { "type": "integer", "minimum": 0 }
        }
      },
...
      "AddressCreateInput": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "address": { "type": "string", "minLength": 1 },
          "ward": { "type": "string" },
          "district": { "type": "string" },
          "province": { "type": "string" },
          "type_address": {
            "type": "string",
            "enum": ["HOME", "WORK", "OTHER"]
          },
          "is_default": { "type": "boolean", "default": false }
        },
        "required": ["user_id", "address", "is_default"]
      },
...
      "AddressResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "minimum": 0, "exclusiveMinimum": true },
          "user_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "address": { "type": "string" },
          "ward": { "type": "string", "nullable": true },
          "district": { "type": "string", "nullable": true },
          "province": { "type": "string", "nullable": true },
          "type_address": {
            "type": "string",
            "nullable": true,
            "enum": ["HOME", "WORK", "OTHER", null]
          },
          "is_default": { "type": "boolean" },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          }
        },
        "required": [
          "id",
          "user_id",
          "address",
          "ward",
          "district",
          "province",
          "type_address",
          "is_default",
          "created_at",
          "updated_at"
        ]
      },
      "CartItemCreateInput": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "product_id": { "type": "string", "format": "uuid" },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": ["user_id", "product_id", "quantity"]
      },
...
      "CartItemResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "minimum": 0, "exclusiveMinimum": true },
          "user_id": { "type": "number" },
          "product_id": { "type": "number" },
          "quantity": { "type": "number" },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          }
        },
        "required": ["id", "user_id", "product_id", "quantity", "created_at", "updated_at"]
      },
      "CategoryCreateInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "description": { "type": "string" },
          "parent_id": { "type": "string", "format": "uuid" },
          "breadcrumbs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CategoryBreadcrumb" },
            "default": []
          }
        },
        "required": ["name", "breadcrumbs"]
      },
...
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "parent_id": { "type": "number", "nullable": true },
          "is_deleted": { "type": "boolean" },
          "breadcrumbs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CategoryBreadcrumb" }
          },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          }
        },
        "required": ["uuid", "name", "description", "parent_id", "is_deleted", "created_at", "updated_at"]
      },
...
      "OrderCreateInput": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "address_id": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderItem" },
            "minItems": 1
          }
        },
        "required": ["user_id", "address_id", "items"]
      },
...
      "OrderResponse": {
        "type": "object",
        "properties": {
          "uuid": { "type": "string", "format": "uuid" },
          "user_id": { "type": "number" },
          "address_id": { "type": "number" },
          "total_amount": { "type": "number" },
          "status": { "type": "string" },
          "is_deleted": { "type": "boolean" },
          "items": { "type": "array", "items": {} },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          },
          "updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "string" }]
          }
        },
        "required": [
          "uuid",
          "user_id",
          "address_id",
          "total_amount",
          "status",
          "is_deleted",
          "created_at",
          "updated_at"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "product_id": { "type": "string", "format": "uuid" },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": ["product_id", "quantity"]
      },
...
      "ReviewCreateInput": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "product_id": { "type": "string", "format": "uuid" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string" }
        },
        "required": ["user_id", "product_id", "rating"]
      },
...
      "ReviewFilter": {
        "type": "object",
        "properties": {
          "product_id": { "type": "string", "format": "uuid" },
          "min_rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "max_rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "limit": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "offset": { "type": "integer", "minimum": 0 }
        }
      },
...
    }
  }
}